amry调整

+++ Assets/Interface/ISkill.cs
@@ -7,5 +7,5 @@ using UnityEngine.UI;
 public interface ISkill
 {
     List<Skill> Skills { get; }
-    void SetSkill(string skillNameZH, string skillNameEN, string commentChinese, int skillCooling, int skillPre_Swing, int skillPost_Swing);
+    void SetSkill(string skillNameZH, string skillNameEN, string commentChinese, float skillCooling, float skillPre_Swing, float skillPost_Swing);
 }
