+++ Assets/Script/ScriptableObject/OtherBuildingSO.cs
@@ -97,7 +97,7 @@ public class OtherBuildingSO : ScriptableObject, IBaseInfo, IBuilding, IWeapon, 
         this.buildingAndPlacementTime = buildingAndPlacementTime;
         this.powerConsume = powerConsume;
     }
-    public void SetSkill(string skillNameZH, string skillNameEN, string commentChinese, int skillCooling, int skillPre_Swing, int skillPost_Swing)
+    public void SetSkill(string skillNameZH, string skillNameEN, string commentChinese, float skillCooling, float skillPre_Swing, float skillPost_Swing)
     {
         if (skills == null)
             skills = new List<Skill>();
